fix deprecation warning: plugins/editor

+++ spyder/plugins/editor/widgets/recover.py
@@ -227,8 +227,8 @@ class RecoveryDialog(QDialog):
     def center(self):
         """Center the dialog."""
         screen = QApplication.desktop().screenGeometry(0)
-        x = screen.center().x() - self.width() / 2
-        y = screen.center().y() - self.height() / 2
+        x = int(screen.center().x() - self.width() / 2)
+        y = int(screen.center().y() - self.height() / 2)
         self.move(x, y)
 
     def restore(self, idx):
